unified _typing
