+++ src/transformers/generation/logits_process.py
@@ -20,9 +20,9 @@ from typing import TYPE_CHECKING, Any, cast
 import numpy as np
 import torch
 
+from .._typing import WhisperGenerationConfigLike
 from ..utils import add_start_docstrings
 from ..utils.logging import get_logger
-from ._typing import WhisperGenerationConfigLike
 
 
 # TODO (joao): We shouldn't need this, but there would be a circular import
